requirements.txt
alembic==1.13.2
asttokens==2.4.1
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jwt==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.32
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.0.3
